requirements.txt
# LangChain AI Agent - Unified Requirements
# Compatible with all branches: openAI, googleAIStudio, llama3.2, main

# === CORE LANGCHAIN (Required for all branches) ===
langchain>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0

# === BRANCH-SPECIFIC PACKAGES ===

# OpenAI branch (Azure OpenAI)
langchain-openai>=0.2.0

# GoogleAI Studio branch 
google-generativeai>=0.8.0

# Llama3.2/Main branches (Ollama)
langchain-community>=0.3.0

# === FASTAPI WEB SERVICE ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# === COMMON DEPENDENCIES ===
# These are automatically installed with the above packages
requests>=2.18.0
pyyaml>=5.3.0
typing-extensions>=4.7.0

# === OPTIONAL DEVELOPMENT TOOLS ===
# Uncomment if needed for testing/development
# pytest>=7.0.0
# jupyter>=1.0.0